+++ app/__init__.py
@@ -2,5 +2,5 @@
 
 from __future__ import annotations
 
-APP_VERSION = "220926.0"
+APP_VERSION = "230926.0"
 __version__ = APP_VERSION